Exported NordVPN class in external module
